Split compiler/rustc_apfloat/tests/ieee.rs
